backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
orjson>=3.9.0
redis>=5.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
httpx>=0.26.0
numpy>=1.26.0
google-generativeai>=0.8.0
